fix(i2c): moving API to use bytes instead of ints where appropriate

+++ src/main/c/i2c.c
@@ -45,7 +45,7 @@ int Java_org_arl_jhwbus_I2CDevice_I2COpen(JNIEnv* env, jobject obj, jstring dev,
     return fd;
 }
 
-int Java_org_arl_jhwbus_I2CDevice_I2CSetAddr(JNIEnv* env, jobject obj, jint fd, jint addr){
+int Java_org_arl_jhwbus_I2CDevice_I2CSetAddr(JNIEnv* env, jobject obj, jint fd, jbyte addr){
     log_info("Setting I2C addr 0x%02X", addr);
     int rv = 0;
     #ifdef __linux__
@@ -63,7 +63,7 @@ int Java_org_arl_jhwbus_I2CDevice_I2CReadByte(JNIEnv* env, jobject obj, jint fd)
     return rv;
 }
 
-int Java_org_arl_jhwbus_I2CDevice_I2CWriteByte(JNIEnv* env, jobject obj, jint fd, jint data){
+int Java_org_arl_jhwbus_I2CDevice_I2CWriteByte(JNIEnv* env, jobject obj, jint fd, jbyte data){
     int rv = 0;
     log_info("Writing a byte 0x%02X", data);
     #ifdef __linux__
@@ -72,7 +72,7 @@ int Java_org_arl_jhwbus_I2CDevice_I2CWriteByte(JNIEnv* env, jobject obj, jint fd
     return rv;
 }
 
-int Java_org_arl_jhwbus_I2CDevice_I2CWriteByteData(JNIEnv* env, jobject obj, jint fd, jint cmd, jint data){
+int Java_org_arl_jhwbus_I2CDevice_I2CWriteByteData(JNIEnv* env, jobject obj, jint fd, jbyte cmd, jbyte data){
     int rv = 0;
     log_info("Writing a command 0x%02X and byte data 0x%02X", cmd, data);
     #ifdef __linux__
@@ -81,7 +81,7 @@ int Java_org_arl_jhwbus_I2CDevice_I2CWriteByteData(JNIEnv* env, jobject obj, jin
     return rv;
 }
 
-int Java_org_arl_jhwbus_I2CDevice_I2CWriteWordData(JNIEnv* env, jobject obj, jint fd, jint cmd, jint data){
+int Java_org_arl_jhwbus_I2CDevice_I2CWriteWordData(JNIEnv* env, jobject obj, jint fd, jbyte cmd, jint data){
     int rv = 0;
     log_info("Writing a command 0x%02X and word data 0x%04X", cmd, data);
     #ifdef __linux__
@@ -90,7 +90,7 @@ int Java_org_arl_jhwbus_I2CDevice_I2CWriteWordData(JNIEnv* env, jobject obj, jin
     return rv;
 }
 
-int Java_org_arl_jhwbus_I2CDevice_I2CReadByteData(JNIEnv* env, jobject obj, jint fd, jint cmd){
+int Java_org_arl_jhwbus_I2CDevice_I2CReadByteData(JNIEnv* env, jobject obj, jint fd, jbyte cmd){
     int rv = 0;
     log_info("Reading a byte data with the command 0x%02X", cmd);
     #ifdef __linux__
@@ -99,7 +99,7 @@ int Java_org_arl_jhwbus_I2CDevice_I2CReadByteData(JNIEnv* env, jobject obj, jint
     return rv;
 }
 
-int Java_org_arl_jhwbus_I2CDevice_I2CReadWordData(JNIEnv* env, jobject obj, jint fd, jint cmd){
+int Java_org_arl_jhwbus_I2CDevice_I2CReadWordData(JNIEnv* env, jobject obj, jint fd, jbyte cmd){
     int rv = 0;
     log_info("Reading a word data with the command 0x%02X", cmd);
     #ifdef __linux__
